requirements.txt
pip

Django==2.2.5
requests==2.22.0
djangorestframework==3.10.3
drf-yasg==1.15.1